+++ servicebus/data-plane/azure-servicebus/src/test/java/com/microsoft/azure/servicebus/SendReceiveTests.java
@@ -88,6 +88,9 @@ public abstract class SendReceiveTests extends Tests {
 
     @AfterClass
     public static void cleanupAfterAllTest() throws ExecutionException, InterruptedException, IOException {
+        if (managementClient == null) {
+            return;
+        }
         if(SendReceiveTests.entityNameCreatedForAllTests != null) {
             managementClient.deleteQueueAsync(SendReceiveTests.entityNameCreatedForAllTests).get();
             managementClient.close();
